+++ stelvio/aws/dns.py
@@ -1,58 +1,23 @@
 import pulumi_aws
-from pulumi import Input, Output, ResourceOptions
+from pulumi import Input, ResourceOptions
 
-from stelvio import dns
-
-
-class Route53PulumiResourceAdapter(dns.Record):
-    @property
-    def name(self) -> Output[str]:
-        return self.pulumi_resource.name
-
-    @property
-    def type(self) -> Output[str]:
-        return self.pulumi_resource.type
-
-    @property
-    def value(self) -> Output[str]:
-        return self.pulumi_resource.records.apply(lambda records: records[0])
+from stelvio.dns import Record
 
 
-class Route53Dns(dns.Dns):
+class Route53Dns:
     def __init__(self, zone_id: str):
         self.zone_id = zone_id
-
-    def create_caa_record(  # noqa: PLR0913
-        self,
-        resource_name: str,
-        name: str,
-        record_type: str,
-        content: str,
-        ttl: int = 1,
-        *,
-        opts: ResourceOptions | None = None,
-    ) -> dns.Record:
-        validation_record = pulumi_aws.route53.Record(
-            resource_name,
-            zone_id=self.zone_id,
-            name=name,
-            type=record_type,
-            records=[content],
-            ttl=ttl,
-            opts=opts,
-        )
-        return Route53PulumiResourceAdapter(validation_record)
 
     def create_record(  # noqa: PLR0913
         self,
         resource_name: str,
-        name: str,
-        record_type: str,
+        name: Input[str],
+        record_type: Input[str],
         value: Input[str],
         ttl: int = 1,
         *,
         opts: ResourceOptions | None = None,
-    ) -> dns.Record:
+    ) -> Record:
         record = pulumi_aws.route53.Record(
             resource_name,
             zone_id=self.zone_id,
@@ -62,4 +27,4 @@ class Route53Dns(dns.Dns):
             ttl=ttl,
             opts=opts,
         )
-        return Route53PulumiResourceAdapter(record)
+        return Record(record, record.name)